Move Event out of fx package
