+++ src/test/kotlin/me/nekomatamune/ygomaker/fx/EventSpec.kt
@@ -1,5 +1,6 @@
 package me.nekomatamune.ygomaker.fx
 
+import me.nekomatamune.ygomaker.*
 import org.spekframework.spek2.Spek
 import strikt.api.expectThat
 import strikt.assertions.isEqualTo
@@ -7,7 +8,8 @@ import strikt.assertions.isEqualTo
 object EventSpec : Spek({
 	group("EventDispatcher#dispatchEvent") {
 		var invocations = 0
-		registerEventHandler(EventName.LOAD_PACK) {
+		registerEventHandler(
+			EventName.LOAD_PACK) {
 			invocations++
 			Result.success(Unit)
 		}
@@ -17,15 +19,18 @@ object EventSpec : Spek({
 		}
 
 		test("Should invoke event handlers multiple times") {
-			dispatchEvent(Event(name = EventName.LOAD_PACK))
+			dispatchEvent(Event(
+				name = EventName.LOAD_PACK))
 			expectThat(invocations).isEqualTo(1)
 
-			dispatchEvent(Event(name = EventName.LOAD_PACK))
+			dispatchEvent(Event(
+				name = EventName.LOAD_PACK))
 			expectThat(invocations).isEqualTo(2)
 		}
 
 		test("Should not invoke handler with non-matching EventName") {
-			dispatchEvent(Event(name = EventName.SAVE_PACK))
+			dispatchEvent(Event(
+				name = EventName.SAVE_PACK))
 			expectThat(invocations).isEqualTo(0)
 		}
